+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to the "cph-ng" extension will be documented in this file.
 
 - **fix**: ignore empty request data in companion
 - **fix**: show or hide compilation channel based on message content
+- **feat**: add file toggle
 
 ## 0.1.8