+++ CHANGELOG.md
@@ -8,6 +8,8 @@ All notable changes to the "cph-ng" extension will be documented in this file.
 - **fix**: show or hide compilation channel based on message content
 - **feat**: add file toggle
 - **fix**: improve folder selection display in quick pick
+- **feat**: add showPanel setting to control visibility of the problem source
+  file panel
 
 ## 0.1.8