+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to the "cph-ng" extension will be documented in this file.
 
-## [Unreleased]
+## 0.1.9
 
 - **fix**: ignore empty request data in companion
 - **fix**: show or hide compilation channel based on message content